Workaround to avoid incompatibility
This changes the Valve OpenVR plugin to always return true for the
method used to determine whether or not to use steamvr input. This is
necessary because I found that there is at least one incompatibility
with another BepInEx plugin name ConfigurationManager. It isn't clear
what the root cause is, but calling GetTypes, as is done in
OpenVRHelpers.cs, causes an exception to be thrown when this plugin is
loaded first. Since for this mod we want this to always be true, I'll
just hard code it instead to avoid the exception. It would be best to
understand the root cause of why the error is thrown, but this seems to
at least prevent the crash.