Add ability to recenter tracking using Home key
If you press the Home key, the tracking is recentered. As an aside, I
noticed while testing this that recentering resulted in my shifting away
from the head position of the player. This is probably the root cause of
why I have been having issues finding the correct center location of the
player head for first person mode. I probably either need to get the
tracking data of the headset and center the camera relative to that or I
need to recenter right before finding the head position. Calculating the
offset based on current tracking position seems better since there might
be side effects of recentering.