More reliable head positioning
This updates the positioning of the head to account for the HMD world
position and should resolve problems of players having their head
position be offset from the character body. Leaving the ability for the
player to adjust manually if they need or want a slightly different
position for preference.