Dont pause or change render scale on dashboard
This patch skips a SteamVR_Render function that is called when the
dashboard gains or loses focus. The function is responsible for pausing
the game when the dashboard is visible as well as setting render scale
to 0.5. Pausing the game breaks the game completely, so when the game is
unpaused, things do not get restored.

Setting the renderscale to half doesn't break the game completely, but
things still end up being blurry even after the render scale is
restored, so something doesn't work right there.

Skipping it doesn't cause any problems and fixes a major problem users
have.

fixes #4