Add crosshair support
This change adds support for the crosshairs. What you interact with and
aiming is tied to the direction you are looking, so this change places
all the relevant crosshair and interaction info into a canvas directly
in front of you.

This is not an ideal solution for the crosshair and it needs work. The
text that is displayed for interacting with objects is alright, but the
normal aiming reticle and the bow aimer, while in the center of your
vision, are off when it comes to where you are actually aiming due to
how the eyes are focusing on objects with stereoscopic displays.

I'm including it in the next release since, while flawed, is an
improvement over the current crosshair, which just sits in the middle of
the GUI.

This should be replaced with a reticlule that is displayed at the depth
the object being aimed at is actually at. Eventually with motion
controller support, the aiming reticle can be attached directly to the
bow and aiming can happen more naturally.