Fix bug with broken crosshair after death
It turns out simply moving the crosshair elements off of the man GUI
canvas onto the separate canvas broke things after the player died. All
of the crosshair elements are normally destroyed and rebuilt on death &
respawn, but because I removed the elements from the canvas, they were
not getting recreated correctly.

Instead of moving the crosshairs off of the canvas, I am just cloning
them, adding the clones to the separate canvas, and disabling the main
canvas crosshair elements.

This change also moves crosshair management into its own class as there
will need to be more logic to deal with moving the crosshairs to the
proper depth and scaling them.

fixes #16