Account for depth in crosshair positioning
This change causes the crosshair to be rendered at the depth that the
object being aimed at exists as well as scales it appropriately to keep
it at the same draw size on screen.

This allows the crosshair to actually be effective in aiming with ranged
weapons and prevents double images from forming since you will no longer
have to focus on the crosshair and an object at a different depth
simultaneously.

resolves #15