Do not set VRPlayer as UIPanel parent
The UI Panel uses a non-convex MeshCollider. After parenting the UI
Panel to the VRPlayer, and subsequently parenting the VRPlayer to the
game's Player model, an error is thrown related to using a non-convex
mesh collider with RigidBodies, which is not allowed.

This was causing the laser pointers to no longer get a raycast hit with
the UI panel once attached to the Player.

The position and rotation are now updated manually each frame.

fixes #18