Complete EnemyHuds impelmentation
This change completes implementing EnemyHuds, so other characters health
bars etc should be visible for a time after the player vision hovers
over them. It is using a Transpiler to inject method calls to my own
class that is being used to create on-demand world-space Canvases that
will be located at the character with the Hud. The transpiler injects
method calls to the UpdateHuds method to update my copy of the hud data
to make sure it is getting updated appropriately.