Update First Person camera positioning
Previous to this change, the first person camera position was being tied
to the player model position using the current HMD transform every
cycle. This turned out to be a bad idea because it means if the player
leans forward, the positioning of the camera is moved backwards to
compensate and keep it on the player model, which creates the effect that
the whole world is moving forwards with your head rather than you moving
your head.

Instead, now it will only account for the current HMD position on first
startup or after HMD tracking is recentered. The offset calculated is
saved and used as a constant offset until next recenter. This change
also removes the usage of the head repositioning config options for
first person modeas it no longer makes sense to use them given this new
strategy. The player can still adjust them in game, but everytime the
tracking is recentered this offset will be reset to zero.