Fix crosshair and layer related things
This commit bundles a few updates. First the crosshair position
management was fixed. It was pushing all the elements in the crosshair
right to the center of the crosshair canvas. Now rather than doing that
I'm just properly parenting things so everything is moved automatically
but in the correct location on the canvas. This fixes the position of
things like the stealth bar and piece health bar.

I also refactored the layer management to a central class since it was
starting to become difficult to keep track of. I combined the crosshair
and enemy hud layers into a central "world space UI" layer and camera
that can be used going forward for any UI that needs to be rendered in
world space outside of the normal UI canvas.