Add placement mode visual cue
In the base game, the placement cursor is centered on the crosshair, so
it is easy to know where the placement cursor is. Since I separated the
cursor from the crosshair, it can be difficult. This change adds a
visual cursor in the form of the Hammer icon to indicate where the
placement cursor is whenever there is not an active placement ghost.

fixes #26