More control tuning and settings
Much closer to being fully featured Index controller supported. Feels
good with all the existing inputs - there is something odd that happens
if moving with the minimap that needs to be fixed.

Had to modify the simulated mouse button presses to more specifically
capture holding the mouse press down rather than having a mouse click
down and release in a single frame. This was required to enable click
and drag to navigate the minimap.