Attach VRIK to PlayerModel
This change moves the VR_IK_Creator into the FinalIK directory so that
it is built with the dll file. It also replaces the bone initialization
with the auto-initializer and sets all the leg bones to null.

The VR_IK_Creator is assigned to the game's player character and assigns
camera and hand targets.