Send UseVrControls option in VRSync
Sending this option to other clients will allow them to respond
appropriately and not add VRIK to other Player when they join if the
other player is not playing with VR controls. This will enable
non-motion controller players and motion controller players to play on
the same server. Non-motion controller players will be able to see the
VR animations of other players and non-motion controller players'
animations will normal to everyone else.