Dont require hands to be active for VRControls
Previously this check was enforcing the the hands are active in order to
activate the VR control action sets. This just resulted in controls
becoming inactive whenever the controller lost tracking, which isn't
helpful at all.

Now as long as the VR Controls option is enabled, the action sets can be
enabled even if the controller loses tracking. That way the button
inputs still work if the player loses controller tracking temporarily.