Fix quick select when hammer equipped
This fixes (albeit in a hacky way) the problem where when the Hammer
equips there is a conflict with the Right Click binding and the
QuickSelect binding. Now when in place mode, QuickSelect will also be
activated by the right click input, so the QuickSelect menu still works
and the weapon can be changed normally without needing to unequip the
hammer in another way first.

fixes #54