Use actual time delta to control rotation speed
Previously, the alt piece rotation speed was being based on a fixed
number of updates occuring, essentially tying the speed to framerate.
This fixes that and makes it based on actual time delta instead as well
as exposes a config to allow the user to tweak the delay if they want.