Use publicized assemblies
This change updates the reference to the assembly_valheim.dll to a
publicized version with open access to all private members. This is a
performance optimization step since FieldRefAccess proved to be
expensive.

This requires that the publicized dll be generated.

The change also enables Running of "Unsafe Code Blocks" in the csproj
build file, which is required to read private fields after compiling
against the public version of the assembly.