Add nonVR mode build
Work in progress.

This adds a build flavor that forces nonVR mode to be on always. This
allows for a separate companion mod to be installed by users who want to
play with VR players that is easier to install. The goal is to be able
to distribute a simple DLL without any other dependencies for non-VR
players, enabling easier install, including with a mod manager like
Vortex. This version doesn't work yet though.