Move HealthPanel to new VRHUD
This change adds a new VR HUD which is used to display the healthbar.
Users can optionally select between a camera locked HUD or a wrist-watch
based hud. Options are included to reposition and scale it.

This change only moves the health panel, but future changes will move
other HUD components, such as stamina and GPower indicators.