Disable RectMask2D for Minimap HUD
This was causing the PlayerMarker icon to vanish whenever the Minimap
was moved from one VRHud location to another. I haven't found a way to
mutate the RectMask2D in such a way that it works at different
locations, and disabling it seemed to work without any side effects, so
it is disabled for now.