Fix minimap pin rendering
This change caues the pins to now be rendered properly onto the VR Hud
minimap. It ensures the pins, when instantiated, are moved to the proper
world space coordinates so they appear at the location of the VR HUD
canvas.

It was also required to use the "map" as the parent for the pins instead
of the "pinRoot", but I haven't figured out why. It might be worth
figuring out what the problem was and fixing it properly rather than
using map as the parent of the pins, but this works as is without any
apparent side effects.