Disable room scale movement when attached to seat, saddle, etc. (#151)
Do not use room scale movement when the character is sitting. This will allow free head locotion when sitting instead of forcing the torso to be upright on the chair. One example usage of this is that the player can move the line of sight around obstacles on the ship to observe the surroundings when steering a ship.