Map input toggles minimap with VRHud
This change causes the Map input to toggle the minimap on and off
instead of toggling the "big" map. The big map will be exclusively
toggled by the Quick action menu.

This only applies when using the updated VRHud for the minimap. If
legacy minimap is used, it will use the original inputs.