Move huds to controller transform
Originally I put the canvases on the wrist transform to keep it tied to
the actual model. I was doing this because I was trying to position it
directly on the wrist, and equipping weapons moved the wrist around
which would cause the canvas to no longer be on the wrist. After
feedback and some reflection, its actually preferrable to just keep it
always tied to directly where the controller transform is and it's fine
if the character model wrist becomes detached from the position when
eqipping or unequipping weapons.

I also added some more appropriate defaults for these transforms.