Add version tracker to config
Added version data to the config file so that we can detect when a new
version of the mod is being started up for the first time in a user's
mod install. This lets us do any one time setup that we need to do, if
any, during an installation. Also updated version to 0.9.0.