Bhaptics vest integration (#184)
Add Bhaptics support integration

BhapticsTactsuit.cs is the vibration patterns player, sending to pc bhaptics client BhapticsPatches.cs are all the harmony hooks a folder called bHaptics contains all the tact files needed for the vest. They are the vibration patterns the bhaptics library is added with all the tact files in the make-release.cmd

Functionalities:
The tactsuit only initializes in vr mode and is disabled by default in the config file. On different events, vibration patterns are thrown that can be a single effect or a repeating effect. All the functionalities are in the harmony hooks in BhapticsPatches.cs.

Author: @Astienth