Only call Unity APIs on main thread (#212)
* Only call Unity APIs on main thread

* Only call Unity APIs on main thread

Some crashes to desktop were observed and traced down to Bow Manager
logic. After analysis the most likely cause was calling Unity APIs and
instantiating GameObjects off the main thread. This change moves that
part of the Async function so that it runs on the main thread after the
expensive computation is done.