Workaround for outline bugs.
For some reason, destroying the outline close to the time of disabling a meshCooldown will cause that outline to linger indefinitely;
On the other hand, keeping the outline instance will cause the outline to appear when another unrelated outline is being displayed.
This workaround offers a toggle on whether the outline instance should be kept based on its usage. It will fix most of the outstanding outline bugs, but is not sufficient to stop player outline from being **temporarily** displayed when hitting a target.

The proper fix will involve looking into Outline.cs to see how outlines are cached.