Disable Pause when using overlay gui
Pausing the game sets the timeScale to 0, and this causes the overlay to
not render or function properly. Until a real fix can be found, pausing
will just be disabled if this option is used. I've also changed the
default to false and included a warning for the option as well as a log
statement to indicate that pausing is disabled.