Update for0.217 (#348)
* Pause vrik and vr camera attach when barbering

* Do not force character to face look direction during barbering

During barbering, the look direction is the opposite of the character facing. Trying to force the two to be the same would cause the character to flicker back and forth.

* Remove HeightMapRenderPatch

Remove this patch since 1) HeightMap#Render() no longer exists and 2) this patch is no longer needed for terrain rendering.

* Remove hasPlacementGhost() check when checking alt piece rotatin mode

hasPlacementGhost() is longer a reliable way to check whether the player is in building mode. Removing it fixes the bug that building pieces cannot be rotated.

* Prevent quick-stack-all action from being triggered accidentally

* Update ValheimGameLibs version number