Do not overwrite configs with command line options
This change changes the behavior that caused overriding an option on the
command line to update the config file value on disk. This has tended to
result in user confusion. Instead the command line options are just
stored in runtime memory and used only for the session the command line
argument was provided. If the command line argument is removed, it will
revert back to using whatever is in the config file.

Additionally it will now print out an informational message when an
option is being overriden as there have been several cases where the
user has forgotten the option was added and had issues with the mod not
working (e.g. they added -nonVrPlayer=true to their arguments in Steam
and can't figure out why the mod isn't starting in VR when they return
some time later). This will help us diagnose this situation easily.