Do not destroy bolt before crossbow attack
Bolt should be destroyed after attack and only if the attack succeeds. There's a brief time window after loading a crossbow during which attack is not available, if we destroy the bolt when the player attempts to attack, not only would the attack fail, the crossbow would be reset and the player would have to load it again.