Delay starting VR until scene loaded
This change delays initializing VR code until the "start" scene is
loaded, which signals that the game is actually at the menu screen. The
"loading" scene that appears on screen isn't captured by the UI
management, so while the game is loading the user is awkardly put in an
empty Unity environment.